pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autoshell"
version = "0.1.0"
description = "An AI-powered shell assistant"
readme = "README.md"
requires-python = ">=3.7"
license = {file = "LICENSE"}
authors = [
  {name = "Sriram Srikanth", email = "sriramsrikanth2005@gmail.com"}
]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = [
    "annotated-types>=0.7.0",
    "blessed>=1.20.0",
    "cachetools>=5.5.2",
    "certifi>=2025.1.31",
    "charset-normalizer>=3.4.1",
    "editor>=1.6.6",
    "google-ai-generativelanguage>=0.6.15",
    "google-api-core>=2.24.2",
    "google-api-python-client>=2.163.0",
    "google-auth>=2.38.0",
    "google-auth-httplib2>=0.2.0",
    "google-generativeai>=0.8.4",
    "googleapis-common-protos>=1.69.1",
    "grpcio>=1.71.0",
    "grpcio-status>=1.71.0",
    "httplib2>=0.22.0",
    "idna>=3.10",
    "inquirer>=3.4.0",
    "markdown-it-py>=3.0.0",
    "mdurl>=0.1.2",
    "prompt_toolkit>=3.0.50",
    "proto-plus>=1.26.1",
    "protobuf>=5.29.3",
    "pyasn1>=0.6.1",
    "pyasn1_modules>=0.4.1",
    "pydantic>=2.10.6",
    "pydantic_core>=2.27.2",
    "Pygments>=2.19.1",
    "pyparsing>=3.2.1",
    "readchar>=4.2.1",
    "requests>=2.32.3",
    "rich>=13.9.4",
    "rsa>=4.9",
    "runs>=1.2.2",
    "six>=1.17.0",
    "tqdm>=4.67.1",
    "typing_extensions>=4.12.2",
    "uritemplate>=4.1.1",
    "urllib3>=2.3.0",
    "wcwidth>=0.2.13",
    "xmod>=1.8.1"
]

[project.urls]
"Homepage" = "https://github.com/Humanoid2005/AutoShell"

[project.scripts]
autoshell = "autoshell.scripts.cli:main"

[tool.setuptools]
packages = ["autoshell", "autoshell.scripts"]